presque poupoupidou_bw

+++ include/extract_bloc.h
@@ -4,12 +4,12 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
-#include <decode_entete.h>
+#include "decode_entete.h"
 
 void init_blocks(struct data *d);
 
-int read_bit(BYTE byte, int num_bit);
+int8_t read_bit(BYTE byte, int8_t num_bit);
 
-int16_t * decode_ac_dc(struct data *d, int index, int table_type,FILE* file,  int16_t * block);
+int16_t * decode_ac_dc(struct data *d, int8_t index, int8_t table_type,FILE* file,  int16_t * block);
 
 #endif /* _EXTRACT_BLOC_H_ */
