+++ include/iDCT.h
@@ -2,6 +2,6 @@
 #define _IDCT_H_
 #include <stdint.h>
 
-uint8_t ** iDCT(int **phi);
+uint8_t ** iDCT(int16_t **phi);
 
 #endif /*_IDCT_H_*/
