+++ include/quantification_inverse.h
@@ -1,8 +1,9 @@
 #ifndef _QUANTIFICATION_INVERSE_H_
 #define _QUANTIFICATION_INVERSE_H_
 
-#include <decode_entete.h>
 
-int * quantification_inverse(struct data *d, int index,  const int16_t *frequentiel);
+#include "decode_entete.h"
+
+int16_t * quantification_inverse(struct data *d, int8_t index,  const int16_t *frequentiel);
 
 #endif /* _QUANTIFICATION_INVERSE_H_ */
