+++ include/zig_zag.h
@@ -1,6 +1,6 @@
 #ifndef _ZIG_ZAG_H_
 #define _ZIG_ZAG_H_
 
-int ** zig_zag(const int *tab);
+int16_t ** zig_zag(const int16_t *tab);
 
 #endif /* _ZIG_ZAG_H_ */
